feat: Add buttons to edit and delete todo items in GetTodoItem view
The code changes in the `GetTodoItem.cshtml` file add buttons to the view for editing and deleting todo items. These buttons provide convenient links to the corresponding actions in the `TodoItems` controller, allowing users to easily modify or remove individual todo items. This feature enhances the functionality of the GetTodoItem view and improves the user experience by providing direct access to important actions.